MIDI-in support, UI piano keyboard & crash fix
MIDI-in works; if a song header is already read, it becomes enabled. It assumes the MIDI-in device is #1 and that the voice table has instrument 48 which is a multi. I was just testing.
UI now has a keyboard which will show the notes that the first track is playing.
Voice table saver now removes invalid characters from the file name (since it doesn't ask for one yet.)