Config changes
Music playlists are now restricted to decimal keys instead of both decimal and hexadecimal ones. This is to prevent a bypass of having song 5 and song 0x5 both be defined in the same playlist, and it's also nicer to look at decimal indexes.
I added Mario Kart: Super Circuit to the config file.
I added the Pokémon Fire Red and Leaf Green soundtrack to the config file.