Command system changes
Each command is a subset of an interface. Now the command arguments' names will be displayed in the track editor when you go to edit them.
"MODT 0" -> "MOD Type" "Vibrate"
"Arg1" -> "Type"

This also makes it so all events can only be changed to have valid values that the M4A engine would allow.

Certain events have had their data type changed as well.

Tempo in the SongPlayer is now a byte as well, so only valid values could ever be set to it.